refactor(cli): consolidate orchestration layer under pipeline module
Previously the orchestration layer was split across two files with an
awkward boundary: `lib.rs` owned `FileTask` / `FileOutcome` / `run`
while `pipeline.rs` owned `ProcessingReport` / `process_file`. `run`
was a one-line wrapper around `process_file`, so the orchestration
responsibility leaked across both files.

Move `FileTask`, `FileOutcome`, `run`, and the `FileTask` unit tests
into `pipeline.rs`. `lib.rs` is now a thin facade: it declares modules,
re-exports the public API (`FileTask`, `FileOutcome`, `ProcessingReport`,
`run`, `parse_args`, `CliError`), and exposes `for_testing` for
integration tests.

No behavior change; 62 tests pass.
